add methods and unit tests

+++ neo/Network/RPC/Helper.cs
@@ -93,7 +93,7 @@ namespace Neo.Network.RPC
             BigInteger numerator = (1 - ((bits[3] >> 30) & 2)) *
                                    unchecked(((BigInteger)(uint)bits[2] << 64) |
                                              ((BigInteger)(uint)bits[1] << 32) |
-                                              (BigInteger)(uint)bits[0]);
+                                              (uint)bits[0]);
             BigInteger denominator = BigInteger.Pow(10, (bits[3] >> 16) & 0xff);
             return (numerator, denominator);
         }
